+++ src/lib/Select.svelte
@@ -87,6 +87,7 @@
     export let VirtualList = null;
     export let ChevronIcon = null;
     export let ClearIcon = null;
+    export let LoadingIcon = null;
     export let debounce = _debounce;
 
     function filterMethod(args) {
@@ -778,32 +779,6 @@
         color: var(--indicatorColor, --iconsColor);
     }
 
-    .spinner {
-        width: var(--spinnerWidth, 20px);
-        height: var(--spinnerHeight, 20px);
-        color: var(--spinnerColor, #51ce6c);
-        animation: rotate 0.75s linear infinite;
-    }
-
-    .spinner_icon {
-        display: block;
-        height: 100%;
-        transform-origin: center center;
-        width: 100%;
-        position: absolute;
-        top: 0;
-        bottom: 0;
-        left: 0;
-        right: 0;
-        margin: auto;
-        -webkit-transform: none;
-    }
-
-    .spinner_path {
-        stroke-dasharray: 90;
-        stroke-linecap: round;
-    }
-
     .multiSelect {
         display: flex;
         padding: var(--multiSelectPadding, 0 35px 0 16px);
@@ -837,12 +812,6 @@
         overflow: hidden;
         padding: 0px;
         white-space: nowrap;
-    }
-
-    @keyframes rotate {
-        100% {
-            transform: rotate(360deg);
-        }
     }
 </style>
 
@@ -925,19 +894,7 @@
         {/if}
 
         {#if isWaiting}
-            <div class="spinner">
-                <svg class="spinner_icon" viewBox="25 25 50 50">
-                    <circle
-                        class="spinner_path"
-                        cx="50"
-                        cy="50"
-                        r="20"
-                        fill="none"
-                        stroke="currentColor"
-                        stroke-width="5"
-                        stroke-miterlimit="10" />
-                </svg>
-            </div>
+            <svelte:component this={LoadingIcon} />
         {/if}
     </div>